Fix ordering of reads on WeakSeqLock
LOAD-LOAD fences need to be after initially reading the lock count. This is
because we do not want loads made during a read lock to be re-ordered with
reading the lock variable, they must stay after the lock read. This does not
change the checkRead(int) function because we want that lock read to occur
only after reads made during the read lock.